Version 2

+++ src/Point_Guard.java
@@ -1,7 +1,7 @@
 public class Point_Guard {
 
 	private int Velocidad;
-	private int Agilidad;
-	private int Punteria;
+	private float Agilidad;
+	private float Punteria;
 
 }
